Move transaction importer component
